+++ src/App.jsx
@@ -3,8 +3,8 @@ import {Route, Routes} from "react-router-dom";
 
 // Pages
 import HomePage from "./pages/home/HomePage.jsx";
-import BlogsPage from "./pages/blogs/BlogsPage.jsx";
-import BlogDetailsPage from "./pages/blog-details-page/BlogDetailsPage.jsx";
+import BlogsOverviewPage from "./pages/blogs-overview/BlogsOverviewPage.jsx";
+import BlogDetailsPage from "./pages/blog-details/BlogDetailsPage.jsx";
 import NewBlogPage from "./pages/new-blog/NewBlogPage.jsx";
 import NotFoundPage from "./pages/not-found/NotFoundPage.jsx";
 
@@ -19,7 +19,7 @@ function App() {
     <main>
       <Routes>
         <Route path="/" element={<HomePage/>}/>
-        <Route path="/blogs" element={<BlogsPage/>}/>
+        <Route path="/blogs" element={<BlogsOverviewPage/>}/>
         <Route path="/blog/:id" element={<BlogDetailsPage/>}/>
         <Route path="/nieuwe-blog" element={<NewBlogPage/>}/>
         <Route path="*" element={<NotFoundPage/>}/>
